pyproject.toml
[tool.poetry]
name = "test2-parcel-microservice"
version = "0.1.0"
description = ""
authors = ["nkriventsov <61682500+nkriventsov@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
sqlalchemy = "^2.0.36"
uvicorn = "^0.32.0"
pydantic-settings = "^2.6.1"
asyncio = "^3.4.3"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
httpx = "^0.27.2"
celery = "^5.4.0"
redis = "^5.2.0"
kombu = "^5.4.2"
loguru = "^0.7.3"
structlog = "^24.4.0"
aiohttp = "^3.11.10"
fastapi-cache2 = "^0.2.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
coverage = "^7.6.5"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
